refactor: 구조 변경 service 레이어 추가
