+++ src/test/java/lotto/domain/LottoPurchaseManagerTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.service.LottoPurchaseManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
