+++ src/test/java/lotto/domain/LottoResultGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.service.LottoResultGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
